Refactoring unit tests code.
